Abstract LabelScanStoreChaosIT, implemented by Lucene
LuceneLabelScanStoreChaosIT extends abstract LabelScanStoreChaosIT
Other LabelScanStore implementations should also use LabelScanStoreChaosIT
